combined actions,renamed functions,deleted unused components,simplified component structure
