+++ package.json
@@ -15,6 +15,7 @@
   "dependencies": {
     "@ant-design/colors": "^7.2.1",
     "@ant-design/icons": "^5.6.1",
+    "@ant-design/v5-patch-for-react-19": "^1.0.3",
     "@redux-devtools/extension": "^3.3.0",
     "antd": "^5.25.2",
     "react": "^19.1.0",
