+++ src/main.jsx
@@ -2,6 +2,7 @@ import { StrictMode } from "react";
 import { createRoot } from "react-dom/client";
 import { Provider } from "react-redux";
 import store from "./redux/store.js";
+import "@ant-design/v5-patch-for-react-19";
 
 import "./index.css";
 import App from "./App.jsx";
@@ -11,5 +12,5 @@ createRoot(document.getElementById("root")).render(
     <Provider store={store}>
       <App />
     </Provider>
-  </StrictMode>
+  </StrictMode>,
 );
