--- src/redux/actions/addActions.js
@@ -1,6 +1,0 @@
-export const add = (task) => {
-  return {
-    type: "add",
-    payload: task,
-  };
-};
