--- src/redux/actions/changeActions.js
@@ -1,6 +1,0 @@
-export const change = (value) => {
-  return {
-    type: "change",
-    payload: value,
-  };
-};
