--- src/redux/actions/clearInputTaskActions.js
@@ -1,5 +1,0 @@
-export const clearInputTask = () => {
-  return {
-    type: "zero",
-  };
-};
