--- src/redux/actions/crossOutTaskActions.js
@@ -1,6 +1,0 @@
-export const crossOutTask = (id) => {
-  return {
-    type: "crossOut",
-    payload: id,
-  };
-};
