--- src/redux/actions/deleteTaskActions.js
@@ -1,6 +1,0 @@
-export const deleteTask = (id) => {
-  return {
-    type: "delete",
-    payload: id,
-  };
-};
