--- src/redux/actions/editIdTaskActions.js
@@ -1,6 +1,0 @@
-export const edit = (id) => {
-  return {
-    type: "edit",
-    payload: id,
-  };
-};
