--- src/redux/actions/editTaskValueActions.js
@@ -1,6 +1,0 @@
-export const editTaskValue = (value) => {
-  return {
-    type: "editTaskValue",
-    payload: value,
-  };
-};
