--- src/redux/actions/updateTaskActions.js
@@ -1,6 +1,0 @@
-export const updateTask = (id, value) => {
-  return {
-    type: "updateTask",
-    payload: { id, value },
-  };
-};
